base/requirements.txt
atomicwrites==1.4.0
attrs==21.1.0
azure-core==1.13.0
azure-storage-blob==12.8.1
boto3==1.17.67
botocore==1.20.67
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
et-xmlfile==1.1.0
fastparquet==0.5.0
greenlet==1.0.0
idna==2.10
iniconfig==1.1.1
isodate==0.6.0
jmespath==0.10.0
llvmlite==0.36.0
loguru==0.5.3
msrest==0.6.21
numba==0.53.1
numpy==1.20.1
oauthlib==3.1.0
openpyxl==3.0.7
packaging==20.9
pandas==1.2.4
pip==21.0.1
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyarrow==4.0.0
pycparser==2.20
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.4.2
six==1.16.0
SQLAlchemy==1.4.13
thrift==0.13.0
toml==0.10.2
tqdm==4.62.2
urllib3==1.26.4
wheel==0.36.2
wincertstore==0.2
xlrd==2.0.1
